+++ CHANGELOG.md
@@ -14,6 +14,10 @@ This change log follows the format documented in [Keep a CHANGELOG].
 
 - Fixed the inferred coerce function types.
 
+### Added
+
+- Added support for nested types.
+
 ## v0.1.0 - 2024-05-27
 
 Initial version